requirements.txt
uvicorn==0.20.0
fastapi==0.87.0
requests==2.28.2
pydantic==1.10.4
starlette==0.21.0
librosa==0.9.2
opencv-python==4.7.0.68
tensorflow-gpu==2.10.1
pypyr
python-dotenv
gunicorn
scikit-learn
scipy
pandas
matplotlib
tqdm
boto3